\name{extractEstimates}
\alias{extractEstimates}
\alias{extractSummaries}
\title{ Extract estimates and summaries from list of fits }
\description{ Extract estimates and summaries from list of fits. }
\usage{
extractEstimates(fits, n = 5)
extractSummaries(fits)
}
\arguments{
  \item{fits}{ the list of fits. }
     \item{n}{ the number of parameters. }
}
\value{
  a matrix for 'extractEstimates',
  or a list of the summaries for 'extractSummaries'.
}
\author{ Jens Henrik Badsberg }
\keyword{ package }
